pyproject.toml
[project]
name = "fantasy-rag-lab"
version = "0.1.0"
description = "Fantasy RAG Project using Wheel of Time and other series"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn",
    "python-dotenv",
    "langchain",
    "langchain-community",
    "langchain-openai",
    "langchain-text-splitters",
    "langgraph",
    "google-genai",
    "langchain-google-genai",
    "pymilvus",
    "neo4j",
    "pypdf",
    "ebooklib",
    "unstructured",
    "sentence-transformers",
    "requests",
    "openai",
    "tiktoken",
    "pandas",
    "pyarrow",
    "jupyter",
    "notebook",
    "ipykernel",
    "structlog>=25.5.0",
    "pydantic-settings>=2.12.0",
    "langchain-voyageai>=0.3.0",
    "pypandoc>=1.16.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
